refactor: split world.rs into world/mod.rs + autotile.rs + buildings.rs + worldgen.rs
3939-line world.rs split into 5 focused submodules (all under 1100 lines each).
No behavior change; pub use re-exports preserve all existing import paths.
